Move Rego builtins to internal/rego sub-packages
Resolves: #1362
